stacks/requirements.txt
aws-cdk-lib==2.3.0
aws-sam-translator==1.42.0
boto3==1.21.10
botocore==1.24.10
cfn-lint==0.58.2
constructs==10.0.75
jmespath==0.10.0
awscli
cdk-ec2-key-pair
cdk-iam-floyd==0.300.0